Separate grants and certs on cause page
